Presentation/scrapper/requirements.txt
Django==3.1
selenium==3.141.0
elasticsearch==7.9.0
python-decouple==3.3
numpy==1.19.4
whitenoise==5.2.0
djangorestframework==3.12.2
gunicorn==20.0.4
whitenoise==5.2.0